Updated youtube_it dependency. Table prefix on migration. Some default URL parameters.
I should probably of created a DB migration for the prefix change, but I didn't reallky want to add
the clutter for such a simple change.

I submitted a push request to youtube_it to allow for specifying url parameters in the youtube embed.
I'm using the custom version of the gem in my app, and I've included some default URL params to customize
youtube embed theme.